refactor(index): add Italian flag icon separately
Instead of installing the entire flags icon pack (which for some reason
isn't being tree shaken properly and is causing my Vercel functions to
be around 1.91 MB in size), this patch updates the index page to simply
add that single icon as a React component.

Fixes: e1dae960ef76 ("feat(index): replace timeline with "dolce" definition")
