+++ app/routes/_index/route.tsx
@@ -1,0 +1,11 @@
+import { Header } from './header'
+import { Hero } from './hero'
+
+export default function IndexPage() {
+  return (
+    <>
+      <Header />
+      <Hero />
+    </>
+  )
+}
